requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
pytz==2026.1.post1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
